refactor(primitives): move TempoAddress to primitives crate
Moves the TempoAddress bech32m encoding implementation from tempo-alloy
into tempo-primitives, making it available at tempo_alloy::primitives::TempoAddress.
The tempo-alloy address module now re-exports from primitives.

Amp-Thread-ID: https://ampcode.com/threads/T-019cbe94-ca15-759b-89d9-bd89edbb2685
