+++ crates/primitives/src/lib.rs
@@ -6,6 +6,9 @@
 
 pub use alloy_consensus::Header;
 
+pub mod address;
+pub use address::TempoAddress;
+
 pub mod transaction;
 pub use transaction::{
     AASigned, MAX_WEBAUTHN_SIGNATURE_LENGTH, P256_SIGNATURE_LENGTH, SECP256K1_SIGNATURE_LENGTH,
